requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
openpyxl==3.1.2
jinja2==3.1.2
python-multipart==0.0.6
Office365-REST-Python-Client==2.5.3
matplotlib==3.8.2
weasyprint==60.1
aiofiles==23.2.1